+++ CHANGELOG.md
@@ -101,6 +101,49 @@ do this anymore, so function permissions are introduced which will explicitly gr
 a function for a given role. A pre-requisite to adding a function permission is that the role should
 have select permissions to the target table of the function.
 
+### `ltree` comparison operators
+
+Comparison operators on columns with ``ltree``, ``lquery`` or ``ltxtquery`` types are now supported, for searching through data stored in a hierarchical tree-like structure.
+
+See the documentation at `graphql/core/queries/query-filters` more details on the currently supported ``ltree`` operators.
+
+**Example query:** Select ancestors of an `ltree` argument
+
+```
+query {
+  tree (
+    where: {path: {_ancestor: "Tree.Collections.Pictures.Astronomy.Astronauts"}}
+  ) {
+    path
+  }
+}
+```
+
+**Example response:**
+```
+{
+  "data": {
+    "tree": [
+      {
+        "path": "Tree"
+      },
+      {
+        "path": "Tree.Collections"
+      },
+      {
+        "path": "Tree.Collections.Pictures"
+      },
+      {
+        "path": "Tree.Collections.Pictures.Astronomy"
+      },
+      {
+        "path": "Tree.Collections.Pictures.Astronomy.Astronauts"
+      }
+    ]
+  }
+}
+```
+
 ### Breaking changes
 
 - This release contains the [PDV refactor (#4111)](https://github.com/hasura/graphql-engine/pull/4111), a significant rewrite of the internals of the server, which did include some breaking changes:
@@ -151,6 +194,7 @@ have select permissions to the target table of the function.
 - server: support tracking of functions that return a single row (fix #4299)
 - server: reduce memory usage consumption of the schema cache structures, and fix a memory leak
 - server: add source name in livequery logs
+- server: support ltree comparison operators (close #625)
 - server: support parsing JWT from cookie header (fix #2183)
 - console: allow user to cascade Postgres dependencies when dropping Postgres objects (close #5109) (#5248)
 - console: mark inconsistent remote schemas in the UI (close #5093) (#5181)